moved all the Settings files into a new directory
